Filter out projects without active allocation flag
When returning the list of projects the user is a member of, filter out
any projects where this flag is not set to "true". This is necessary
because some systems, e.g. KVM, do not run any enforcement other than
quota + the existence of project roles. We need some way to disable the
project; this effectively removes any users from it (we remove roles for
projects a user is not a member of during authentication), which is
equivalent to disabling.