Fix join_date on user authentication
Set join date in RequiredActionChallenge instead of in processAction.
processAction is only called when a form is submitted, but this is
noninteractive.

if join date is not set, evaluateTriggers sets adds
requiredActionChallenge to the list of required actions.

Otherwise, if the required action is set to default, or if set manually,
Required action challenge will still run. Therefore, this function also
checks the status of join date, and won't overwrite it if set.