add new authenticator to handle linked globus IDs
This authenticator handles the case where a user creates a chameleon
account with a globus identity, then later links that identity to a new
primary.

When authenticating, globus returns a token containing all associated,
validated identities. If one of those identities matches the one linked
to an existing chameleon user, it is safe to update the linked identity
to the new one.

This commit adds an authenticator to be used in the "broker first login"
flow, and will update an existing user, if and only if an exact match is
found for the linked identity.

Since keycloak cannot handle multiple copies of an IdP linked to a user,
we must delete the linked identity, then re-link the new one.
This is done by attaching the found user to the authentication session,
then returning success.

To avoid orphaning accounts, the deletion and linking must happen
atomically. For example, if a confirmation step was required to link,
the deletion could happen without the relinking.

TODO: The correct way to handle this is to attach an "ExistingUserInfo"
object to the authentication session state, and then a later handler
should handle the conflicts. When implemented, this authenticator will
only handle detection/matching, and the delete+link functions should be
moved to the confirmation step.