check for username,email matches in identitySet
This adds checks for matching username and email, not just linked
identity. However, we cannot assume that a match means the user is
authenticated in this case, and an additional confirmation step is
needed.

Therefore, in these cases, we set "ExistingUserInfo" to the
authentication context, and set context.attempted. This delegates
conflict handling to later authenticators in the flow.

A "sane" flow to handle this can be seen in the default "broker first
login" flow, under "handle existing account". It first prompts the user
to confirm account linking, then emails the listed contact email on the
account for confirmation.

TODO:
The case that is currently unhandled is if a user authenticates with a
new globus account, but alreday has one configured. The
verifyEmailAuthenticator must be modified to handle the duplicate IDP entry.