add user attribute mapper for globus IDP (#10)
* add user attribute mapper for globus IDP

During authentication, globus returns a token containing all associated identities.
We store the unique ID of each identity in the user attribute fields, with key named
globus_sub_<uuid>, and value of "linked". While we could put additional info here,
keycloak attributes have a max length of 255 characters.

* add new authenticator to handle linked globus IDs

This authenticator handles the case where a user creates a chameleon
account with a globus identity, then later unlinks an identity and attempts to log in.

This  adds an authenticator to be used in the "broker first login"
flow, and will update an existing user, if and only if an exact match is
found between the new primary identity, and the stored, linked identities.

Since keycloak cannot handle multiple copies of an IdP linked to a user,
we must delete the linked identity, then re-link the new one.
This is done by attaching the found user to the authentication session,
then returning success.

To avoid orphaning accounts, the deletion and linking must happen
atomically. For example, if a confirmation step was required to link,
the deletion could happen without the relinking.